fix build (#26)
