+++ docs/guides/installation.mdx
@@ -4,11 +4,16 @@ sidebar_position: 1
 ---
 
 import CodeBlock from "@theme/CodeBlock"
-import github from "../../src/gen/github.json"
+import Version from "../../src/components/Version"
 import InstallCommand from "../../src/components/install"
 
 <>
-  You can install the latest version - <b> {github.release.tag_name}</b>, by using one of the following methods:
+  You can install the latest version -{" "}
+  <b>
+    {" "}
+    <Version />
+  </b>
+  , by using one of the following methods:
 </>
 
 ## Homebrew
@@ -53,7 +58,10 @@ If you want to install Tailcall with Docker, follow the steps below. Before star
 
 1. Pull the latest Tailcall Docker image using the following command:
 
-   <CodeBlock language="bash">docker pull ghcr.io/tailcallhq/tailcall/tc-server:{github.release.tag_name}</CodeBlock>
+   <CodeBlock language="bash">
+     docker pull ghcr.io/tailcallhq/tailcall/tc-server:
+     <Version />
+   </CodeBlock>
 
    This command fetches the latest version of the Tailcall Docker image from the Docker registry.
 
@@ -61,7 +69,7 @@ If you want to install Tailcall with Docker, follow the steps below. Before star
 
    <CodeBlock language="bash">
      docker run -p 8080:8080 -p 8081:8081 ghcr.io/tailcallhq/tailcall/tc-server:
-     {github.release.tag_name}
+     <Version />
    </CodeBlock>
 
    This command starts the Tailcall server in a Docker container. Similar to the homebrew installation, it exposes a the graphQL endpoint on port 8080.
