+++ docs/intro/quickstart.mdx
@@ -4,12 +4,16 @@ sidebar_position: 1
 label: Quick Start
 ---
 
-import github from "../../src/gen/github.json"
+import Version from "../../src/components/Version"
 
 ## Installation
 
 <>
-  You can install the latest version - <b>{github.release.tag_name}</b>, by using
+  You can install the latest version -{" "}
+  <b>
+    <Version />
+  </b>
+  , by using
   <b> homebrew</b>.
 </>
